DsTemplate Documentation

DsTemplate.Setup.initialize_project — Function

initialize_project(path;
                   git=true, 
                   template=DEFAULT_TEMPLATE,
                   name="",
                   lang="julia",
                   report="tex",
                   force=false)

path (String) eg "MyProject" will make a folder "MyProject" and initialize all other subfolders there.

git (bool) default true. If the folder should be initialized as a git folder

name (String) defaults to the name of the path. Used when creating a main julia module with an uppercased first letter

lang (String) default "julia". Can be "python" too.

report (String) default "tex". Will create a latex file plus bibliography. can also be "qmd" for quarto reports.